master

+++ ft_strlen.c
@@ -6,14 +6,14 @@
 /*   By: vnilprap <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/21 23:28:28 by vnilprap          #+#    #+#             */
-/*   Updated: 2022/03/02 21:21:41 by vnilprap         ###   ########.fr       */
+/*   Updated: 2022/03/07 18:50:33 by vnilprap         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 size_t	ft_strlen(char *c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (c[i])
